refactor: Input class UsersAPIComponent
 Input class UsersAPIComponent into UsersContainer.jsx
